feat(mutator, ui): Implemented Button template, expanded Mutation attribute feature to include methods.
